Improvements according to a review.
