+++ src/test/scala/org/ergoplatform/utils/HistoryTestHelpers.scala
@@ -1,7 +1,7 @@
 package org.ergoplatform.utils
 
 import org.ergoplatform.nodeView.history.ErgoHistory
-import org.ergoplatform.nodeView.history.storage.modifierprocessors.EmptyBlockSectionProcessor
+import org.ergoplatform.nodeView.history.storage.modifierprocessors.{EmptyBlockSectionProcessor, FullBlockPruningProcessor, ToDownloadProcessor}
 import org.ergoplatform.nodeView.state.StateType
 import org.ergoplatform.settings._
 import org.scalacheck.Gen
@@ -57,4 +57,25 @@ trait HistoryTestHelpers extends ErgoPropertyTest {
 
     ErgoHistory.readOrGenerate(fullHistorySettings, timeProvider)
   }
+
+}
+
+object HistoryTestHelpers {
+
+  /**
+    * Use reflection to set `minimalFullBlockHeightVar` to 0 to change regular synchronization rule, that we
+    * first apply headers chain, and apply full blocks only after that
+    */
+  def allowToApplyOldBlocks(history: ErgoHistory): Unit = {
+    import scala.reflect.runtime.{universe => ru}
+    val runtimeMirror = ru.runtimeMirror(getClass.getClassLoader)
+    val procInstance = runtimeMirror.reflect(history.asInstanceOf[ToDownloadProcessor])
+    val ppM = ru.typeOf[ToDownloadProcessor].member(ru.TermName("pruningProcessor")).asMethod
+    val pp = procInstance.reflectMethod(ppM).apply().asInstanceOf[FullBlockPruningProcessor]
+    val f = ru.typeOf[FullBlockPruningProcessor].member(ru.TermName("minimalFullBlockHeightVar")).asTerm.accessed.asTerm
+    runtimeMirror.reflect(pp).reflectField(f).set(ErgoHistory.GenesisHeight)
+    val f2 = ru.typeOf[FullBlockPruningProcessor].member(ru.TermName("isHeadersChainSyncedVar")).asTerm.accessed.asTerm
+    runtimeMirror.reflect(pp).reflectField(f2).set(true)
+  }
+
 }
